+++ src/index.ts
@@ -1,6 +1,3 @@
-import * as walletSdk from "./walletSdk";
-import { Keypair } from "stellar-sdk";
-
 /**
  * Types
  */
@@ -20,6 +17,7 @@ export {
 } from "./walletSdk/Horizon";
 export { Interactive } from "./walletSdk/Interactive";
 export { Recovery } from "./walletSdk/Recovery";
+export { Watcher } from "./walletSdk/Watcher";
 
 /**
  * Utils
@@ -33,6 +31,8 @@ export { Utils };
 import * as Exceptions from "./walletSdk/Exceptions";
 export { Exceptions };
 
+import * as walletSdk from "./walletSdk";
+import { Keypair } from "stellar-sdk";
 // TODO - figure out why Keypair used in parent codebase throws error
 export { walletSdk, Keypair };
 export default { walletSdk };
