+++ src/walletSdk/Types/index.ts
@@ -3,3 +3,4 @@ export * from "./auth";
 export * from "./horizon";
 export * from "./interactive";
 export * from "./utils";
+export * from "./watcher";
