--- src/walletSdk/Watcher/Types.ts
@@ -1,7 +1,0 @@
-export type WatcherRefreshFunction = () => void;
-export type WatcherStopFunction = () => void;
-
-export interface WatcherResponse {
-  refresh: WatcherRefreshFunction;
-  stop: WatcherStopFunction;
-}
